SDI-33 Send message for each visit to be migrated

+++ src/main/kotlin/uk/gov/justice/digital/hmpps/prisonerfromnomismigration/service/NomisApiService.kt
@@ -19,8 +19,8 @@ class NomisApiService(@Qualifier("nomisApiWebClient") private val webClient: Web
     visitTypes: List<String>,
     fromDateTime: LocalDateTime?,
     toDateTime: LocalDateTime?,
-    pageNumber: Int,
-    pageSize: Int,
+    pageNumber: Long,
+    pageSize: Long,
   ): PageImpl<VisitId> =
     webClient.get()
       .uri {
